requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
jsonify
json
requests
werkzeug.exceptions
werkzeug.security
functools